Move Tidal HiFi instances to Integration; manual on-demand test with online/deprecated ping; drop auto-discovery and explanatory blurbs
